[UPDATE] Refactor folder structure
