+++ lib/features/home/models/tips_model.dart
