+++ lib/features/login/repository/login_repository.dart
@@ -2,7 +2,7 @@ import 'package:dio/dio.dart';
 import 'package:loomi_test/support/constants.dart';
 import 'package:retrofit/retrofit.dart';
 
-import '../../models/logged_user_model.dart';
+import '../models/logged_user_model.dart';
 
 part 'login_repository.g.dart';
