+++ test/home/repository/home_repository_test.dart
@@ -1,10 +1,10 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:loomi_test/models/bonus_model.dart';
-import 'package:loomi_test/models/championship_model.dart';
-import 'package:loomi_test/models/match_model.dart';
-import 'package:loomi_test/models/tips_model.dart';
-import 'package:loomi_test/models/won_bets_model.dart';
-import 'package:loomi_test/repositories/home/home_repository.dart';
+import 'package:loomi_test/features/home/models/bonus_model.dart';
+import 'package:loomi_test/features/home/models/championship_model.dart';
+import 'package:loomi_test/features/home/models/match_model.dart';
+import 'package:loomi_test/features/home/models/tips_model.dart';
+import 'package:loomi_test/features/home/models/won_bets_model.dart';
+import 'package:loomi_test/features/home/repository/home_repository.dart';
 import 'package:mocktail/mocktail.dart';
 
 import '../shared/home_shared.dart';
