+++ test/home/shared/home_shared.dart
@@ -1,9 +1,9 @@
-import 'package:loomi_test/models/bonus_model.dart';
-import 'package:loomi_test/models/championship_model.dart';
-import 'package:loomi_test/models/match_model.dart';
-import 'package:loomi_test/models/tips_model.dart';
-import 'package:loomi_test/models/won_bets_model.dart';
-import 'package:loomi_test/repositories/home/home_repository.dart';
+import 'package:loomi_test/features/home/models/bonus_model.dart';
+import 'package:loomi_test/features/home/models/championship_model.dart';
+import 'package:loomi_test/features/home/models/match_model.dart';
+import 'package:loomi_test/features/home/models/tips_model.dart';
+import 'package:loomi_test/features/home/models/won_bets_model.dart';
+import 'package:loomi_test/features/home/repository/home_repository.dart';
 import 'package:mocktail/mocktail.dart';
 
 class HomeShared {
@@ -25,11 +25,11 @@ class HomeShared {
       teamB: "teamB",
       teamAImage: "teamAImage",
       teamBImage: "teamBImage",
-      teamAScore: "teamAScore",
-      teamBScore: "teamBScore",
-      xbet: "xbet",
-      betsafe: "betsafe",
-      betsson: "betsson",
+      teamAScore: 10000,
+      teamBScore: 10000,
+      xbet: 10000,
+      betsafe: 10000,
+      betsson: 10000,
     ),
   ];
