+++ test/login/repository/login_repository_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:loomi_test/repositories/login/login_repository.dart';
+import 'package:loomi_test/features/login/repository/login_repository.dart';
 import 'package:mocktail/mocktail.dart';
 
 import '../shared/login_shared.dart';
